requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
# Keras is now included within TensorFlow for recent versions

# Data Processing & Numerical Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars & Utilities
tqdm>=4.65.0
jupyter>=1.0.0

# Model Saving/Loading
h5py>=3.9.0
joblib>=1.3.0
